project1/src/requirements.txt
backports.shutil-get-terminal-size==1.0.0
coverage==4.5.1
decorator==4.2.1
Django==1.11.11
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-json-rpc==0.7.2
enum34==1.1.6
factory-boy==2.11.1
Faker==0.9.1
funcsigs==1.0.2
ipaddress==1.0.22
ipython==5.5.0
ipython-genutils==0.2.0
jsonrpc==1.2
mock==2.0.0
mysqlclient==1.3.13
pathlib2==2.3.0
pbr==4.3.0
pexpect==4.4.0
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.7.3
python-memcached==1.59
pytz==2018.3
scandir==1.7
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.2
traitlets==4.3.2
wcwidth==0.1.7
